Fix CP Not Auto Snatching Bug
Workaround for a bug in https://api.couchpota.to/eta/<imdbid>
CP uses this API to retrieve movie release dates. The API always returns
"{dvd":0,"theater":0,"bluray":false,"expires":<some-bogus-time-about-11-days-in-the future>}
Because of this, CP will never auto-snatch movies because it can't establish a valid ETA
This workaround tries to extract a valid release-date from the 'released' field of the media info
which seems to be valid but seems to vary in format.
If it fails to convert the date, it uses the expires value.